==> brazil_denv_analysis/03_summary_maps_climate_dependent_parameters.Rmd <==
---
title: "Analysis of DENV in Brazil: climate-dependent parameters"
output:
  html_document:
    toc: yes
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::knit_hooks$set(
  margin1 = function(before, options, envir) {
    if (before) par(mgp = c(1.5, .5, 0), bty = "n", plt = c(.105, .97, .13, .97))
    else NULL
  },
  margin2 = function(before, options, envir) {
    if (before) par(mgp = c(2, .5, 0), bty = "n", plt = c(.105, .97, .13, .97))
    else NULL
  },
  margin3 = function(before, options, envir) {
    if (before) par(mgp = c(1.5, .5, 0), bty = "n", mai = rep(.1, 4))
    else NULL
  }
)

knitr::opts_chunk$set(echo       = TRUE,
                      cache      = TRUE, 
                      margin1    = TRUE,
                      fig.retina = 2,
                      fig.align  = "center")
```

```{css, echo=FALSE}
pre {
max-height: 300px;
overflow-y: auto;
}

pre[class] {
max-height: 100px;
}
```

## Links to local data sets

Here we use the following links to data.

```{r}
RAW_CLIMATE_DATA_FOLDER <- file.path("data/bra2_climate/")
```

Here, we define the file paths to the output. 

```{r}
SUMMARY_PLOTS_FOLDER <- "output/plots/"
SUMMARY_RESULTS_FOLDER <- "output/results/"
```

Change them accordingly if you want to run the script locally on your computer.

## Preambule

This document provides an example of the distribution of the climate-dependent parameters over time. Climate-dependent parameters include mosquito biting rate, mosquito lifespan, mosquito-DENV incubation period, mosquito-to-human transmission probability per bite, human-to-mosquito transmission probability per bite. For each of these parameters, we plot a time series that depends on local climatic conditions. We selected Salvador, Bahia as our example. 

## Packages

The needed packages:

```{r message=FALSE}
library(pacman)
p_load(MVSE)
p_load(tidyverse, RColorBrewer, lubridate, cowplot)
```

## Index P simulation

Here, we perform Index P simulation for Rio Branco, Acre (ID: 120040).

```{r}
# climate data
climate_data <- read.csv(file.path(RAW_CLIMATE_DATA_FOLDER, "BR_mun_120040_climate_entire_region.csv"))

# user-defined model
mun_model <- mvse_model(model_name="mun_model", climate_data=climate_data, model_category="aegypti")

# fit model
mun_fit <- MVSE::fitting(object=mun_model, iter=10^6, warmup=0.3, seed=123, 
                         init=c(rho=1, eta=10, alpha=3), gauJump=c(rho=2, eta=5), samples=10^4, verbose=TRUE)
```

## Climate-dependent parameter figures

First, let's calculate the summary statistics for the distribution of each climate-dependent parameter. 

```{r}
# organize the summary data for each climate-dependent parameter
par_ids <- c("muV", "aV", "phiVH", "phiHV", "gammaV")
par_names <- c("mosq_lifespan", "mosq_biting_rate", "mosq_human_trans_prob", "human_mosq_trans_prob", "mosq_inc_per")
need_to_inv <- c(TRUE, FALSE, FALSE, FALSE, TRUE)
par_data_summary_list <- list()
for (ii in seq_along(par_ids)) {
  par_data <- mun_fit@sim[[par_ids[ii]]] 
  if (need_to_inv[ii]) par_data <- 1/par_data
  par_data <- as.data.frame(par_data)
  par_data_summary <- cbind(time_point=1:nrow(par_data), par_data) %>% 
    select(time_point, dplyr::everything()) %>% 
    gather(key="sim", value="value", 2:ncol(.)) %>% 
    group_by(time_point) %>%
    summarise(mean=mean(value), sd=sd(value), l50=quantile(value, 0.25), u50=quantile(value, 0.75), 
              l90=quantile(value, 0.05), u90=quantile(value, 0.95), .groups="keep") %>% 
    as.data.frame() %>% 
    mutate(par=par_names[ii])
  par_data_summary_list[[ii]] <- par_data_summary
}
par_data_summary <- do.call(rbind, par_data_summary_list)

dates <- mun_fit@sim$indexP[, "date"] %>% 
  as.data.frame() %>% 
  setNames(c("date")) %>% 
  mutate(year=lubridate::year(date), month=lubridate::month(date))
```

Now, let's create time series plots for each climate-dependent parameter. 

```{r}
# climate time series
climate_data <- mun_model@climate_data %>% 
  select(date, T, H) %>% 
  mutate(time_point=1:nrow(.))
climate_data_plot <- ggplot(data=climate_data) +
  theme_bw() +
  geom_line(mapping=aes(x=time_point, y=`T`*70/22, group=1), color='cadetblue4') + 
  geom_line(mapping=aes(x=time_point, y=`H`, group=1), color='magenta') + 
  scale_y_continuous(limits=c(50, 100), breaks=seq(0, 100, 10), name="\nRelative humidity (%)",
                     sec.axis = sec_axis(trans=~.*22/70, name="Temperature (°C)")) +
  scale_x_continuous(breaks=seq(1, nrow(dates), 12*3), labels=dates$year[seq(1, nrow(dates), 12*3)], 
                     limits=c(0, 15*12+2), expand=c(0.01, 0.05)) + 
  theme(legend.position="bottom", legend.title=element_blank(),
        axis.title.y.left=element_text(color='magenta', size=9),
        axis.title.y.right=element_text(color='cadetblue4', size=9), axis.title.x=element_blank())

# index P time series
indexP_data <- MVSE::extract(mun_fit)[["indexP"]] %>%
  gather(key="sim", value="value", 2:ncol(.)) %>%
  group_by(date) %>%
  summarise(mean=mean(value), sd=sd(value), l90=quantile(value, 0.05), u90=quantile(value, 0.95), .groups="keep") %>%
  as.data.frame() %>%
  mutate(time_point=1:nrow(.))
indexP_data_plot <- ggplot(data=indexP_data) +
  geom_ribbon(mapping=aes(x=time_point, ymin=l90, ymax=u90), fill="red", alpha=0.2) +
  geom_line(mapping=aes(x=time_point, y=mean), color="red") +
  scale_y_continuous(breaks=seq(0, 4, 1)) +
  scale_x_continuous(breaks=seq(1, nrow(dates), 12*3), labels=dates$year[seq(1, nrow(dates), 12*3)],
                     limits=c(0, 15*12+2), expand=c(0.01, 0.05)) +
  labs(y="\nIndex P") +
  theme_bw() +
  theme(axis.title.x=element_blank()) +
  theme(axis.title=element_text(size=9)) +
  coord_cartesian(ylim=c(0, 4))

# mosquito lifespan and mosquito extrinsic incubation period
mosq_lifespan_EIP_plot <- ggplot(data=filter(par_data_summary, par %in% c("mosq_lifespan", "mosq_inc_per"))) + 
  geom_ribbon(mapping=aes(x=time_point, ymin=l90, ymax=u90, fill=par), alpha=0.2) + 
  geom_line(mapping=aes(x=time_point, y=mean, color=par)) + 
  scale_fill_manual(breaks=c("mosq_lifespan", "mosq_inc_per"), values=c(brewer.pal(8, "Dark2")[5], brewer.pal(9, "Blues")[7]), 
                    labels=c("Mosquito lifespan", "Mosquito incubation period")) + 
  scale_color_manual(breaks=c("mosq_lifespan", "mosq_inc_per"), values=c(brewer.pal(8, "Dark2")[5], brewer.pal(9, "Blues")[7]), 
                     labels=c("Mosquito lifespan", "Mosquito incubation period")) + 
  scale_y_continuous(breaks=seq(0, 40, 5)) + 
  scale_x_continuous(breaks=seq(1, nrow(dates), 12*3), labels=dates$year[seq(1, nrow(dates), 12*3)], 
                     limits=c(0, 15*12+2), expand=c(0.01, 0.05)) + 
  labs(y="Vector variables\n(days)") + 
  theme_bw() + 
  theme(axis.title.x=element_blank(), axis.text.x=element_blank()) + 
  theme(legend.position=c(0.85, 0.85), legend.title=element_blank(), legend.background=element_blank(), 
        legend.box.background=element_blank(), legend.key.size=unit(0.5, "cm")) + 
  coord_cartesian(ylim=c(0, 30)) + 
  theme(axis.title=element_text(size=9), legend.text=element_text(size=9))

# transmission probability
mosq_trans_prob_plot <- ggplot(data=filter(par_data_summary, par %in% c("mosq_human_trans_prob", "human_mosq_trans_prob"))) + 
  geom_ribbon(mapping=aes(x=time_point, ymin=l90, ymax=u90, fill=par), alpha=0.2) + 
  geom_line(mapping=aes(x=time_point, y=mean, color=par)) + 
  scale_fill_manual(breaks=c("mosq_human_trans_prob", "human_mosq_trans_prob"), values=c(brewer.pal(8, "Dark2")[3], brewer.pal(8, "Dark2")[4]), 
                    labels=c("Mosquito-to-human", "Human-to-mosquito")) + 
  scale_color_manual(breaks=c("mosq_human_trans_prob", "human_mosq_trans_prob"), values=c(brewer.pal(8, "Dark2")[3], brewer.pal(8, "Dark2")[4]), 
                     labels=c("Mosquito-to-human", "Human-to-mosquito")) + 
  scale_y_continuous(breaks=seq(0.7, 1, 0.1), limits=c(0.7, 1)) + 
  scale_x_continuous(breaks=seq(1, nrow(dates), 12*3), labels=dates$year[seq(1, nrow(dates), 12*3)], 
                     limits=c(0, 15*12+2), expand=c(0.01, 0.05)) + 
  labs(y="\nTransmission probability") + 
  theme_bw() + 
  theme(axis.title.x=element_blank()) + 
  theme(legend.position=c(0.85, 0.2), legend.title=element_blank(), legend.background=element_blank(), 
        legend.box.background=element_blank(), legend.key.size=unit(0.5, "cm")) + 
  theme(axis.title=element_text(size=9), legend.text=element_text(size=9))

# mosquit biting rate
mosq_biting_rate_plot <- ggplot(data=filter(par_data_summary, par %in% c("mosq_biting_rate"))) + 
  geom_ribbon(mapping=aes(x=time_point, ymin=l90, ymax=u90), fill=brewer.pal(8, "Dark2")[7], alpha=0.2) + 
  geom_line(mapping=aes(x=time_point, y=mean), color=brewer.pal(8, "Dark2")[7]) + 
  # scale_y_continuous(breaks=seq(0.7, 1, 0.1), limits=c(0.7, 1)) + 
  scale_x_continuous(breaks=seq(1, nrow(dates), 12*3), labels=dates$year[seq(1, nrow(dates), 12*3)], 
                     limits=c(0, 15*12+2), expand=c(0.01, 0.05)) + 
  labs(y="Mosquito biting rate\n(bites/female/day)") + 
  theme_bw() + 
  theme(axis.title.x=element_blank(), axis.text.x=element_blank()) + 
  theme(legend.position=c(0.77, 0.1), legend.title=element_blank(), legend.background=element_blank(), 
        legend.box.background=element_blank(), legend.key.size=unit(0.5, "cm")) + 
  theme(axis.title=element_text(size=9)) 

pA <- cowplot::plot_grid(NULL, climate_data_plot, ncol=1, rel_heights=c(0, 1))
pB <- cowplot::plot_grid(mosq_lifespan_EIP_plot, mosq_biting_rate_plot, mosq_trans_prob_plot, 
                         ncol=1, rel_heights=c(1, 1, 1.1))
pB <- cowplot::plot_grid(pB, NULL, nrow=1, rel_widths=c(1, 0.057))
pC <- cowplot::plot_grid(indexP_data_plot, NULL, nrow=1, rel_widths=c(1, 0.057))
pABC <- cowplot::plot_grid(pA, pB, pC, ncol=1, labels=c("A", "B", "C"), rel_heights=c(1, 3, 1))
# pAB <- cowplot::plot_grid(pA, pB, labels=c("A", "B"), rel_heights=c(1, 3), ncol=1)
scale_factor <- 0.65
pdf(file=file.path(SUMMARY_PLOTS_FOLDER, "climate_dependent_parameters.pdf"),w=8.27, h=11.69*scale_factor)
print(pABC)
a <- dev.off()
```
