==> brazil_denv_analysis/03_summary_maps_typical_year_dynamics.Rmd <==
---
title: "Analysis of DENV in Brazil: DENV incidence typical year dynamics"
output:
  html_document:
    toc: yes
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::knit_hooks$set(
  margin1 = function(before, options, envir) {
    if (before) par(mgp = c(1.5, .5, 0), bty = "n", plt = c(.105, .97, .13, .97))
    else NULL
  },
  margin2 = function(before, options, envir) {
    if (before) par(mgp = c(2, .5, 0), bty = "n", plt = c(.105, .97, .13, .97))
    else NULL
  },
  margin3 = function(before, options, envir) {
    if (before) par(mgp = c(1.5, .5, 0), bty = "n", mai = rep(.1, 4))
    else NULL
  }
)

knitr::opts_chunk$set(echo       = TRUE,
                      cache      = TRUE, 
                      margin1    = TRUE,
                      fig.retina = 2,
                      fig.align  = "center")
```

```{css, echo=FALSE}
pre {
max-height: 300px;
overflow-y: auto;
}

pre[class] {
max-height: 100px;
}
```

## Links to local data sets

Here we use the following links to data.

```{r}
CASE_DATA_FOLDER <- "data/bra2_cases/"
SF_DATA_FOLDER <- "data/bra2_sf/"
POPULATION_DATA_FOLDER <- "data/bra2_population/"
SUMMARY_OUTPUT_DATA_FOLDER <- "output/data/summary_output/"

# raw data 
bra1_path <- file.path(SF_DATA_FOLDER, "bra1_sf_data.rds")
bra2_path <- file.path(SF_DATA_FOLDER, "simplified_bra2_sf_data.rds")
bra2_pop_data_path <- file.path(POPULATION_DATA_FOLDER, "bra2_population_data_2000_2014.rds")
bra2_case_data_path <- file.path(CASE_DATA_FOLDER, "bra_DENV_case_data_2000_2014.rds")

# processed climate summary data
bra2_entire_period_prec_summary_data_path <- file.path(SUMMARY_OUTPUT_DATA_FOLDER, "region_basic_precipitation_summary_data.Rdata")
bra2_log_incidence_prec_corr_data_path <- file.path(SUMMARY_OUTPUT_DATA_FOLDER,
                                                    "region_typical_year_corr_log_incidence_precipitation_data.Rdata")

# processed index P summary data
bra2_indexP_summary_data_path <- file.path(SUMMARY_OUTPUT_DATA_FOLDER, "region_basic_summary_data.Rdata")

# processed case summary data
bra2_log_incidence_typical_year_summary_data_path <- file.path(SUMMARY_OUTPUT_DATA_FOLDER,
                                                               "region_typical_year_log_incidence_data.Rdata")

# processed indexP/case summary data
bra2_typical_year_corr_log_incidence_data_path <- file.path(SUMMARY_OUTPUT_DATA_FOLDER,
                                                            "region_typical_year_corr_log_incidence_data.Rdata")
```

Here, we define the file paths to the output. 

```{r}
SUMMARY_PLOTS_FOLDER <- "output/plots/"
SUMMARY_RESULTS_FOLDER <- "output/results/"
```

Change them accordingly if you want to run the script locally on your computer.

## Preambule

This document compares the Index P and DENV incidence dynamics for a typical year. We calculate the correlation and time delay as well as other summary statistics of the relationship between the two time series. 

## Packages

The needed packages:

```{r message = FALSE}
library(pacman)
p_load(tidyverse, sf, RColorBrewer, magrittr, cowplot, ggrepel)
```

## Useful colour palettes

```{r}
levColors <- rev(brewer.pal(11, "RdYlBu")) # colourblind friendly
circColors2 <- c(brewer.pal(7, "YlGnBu")[1:6], rev(brewer.pal(6, "RdPu")))
circColors2 <- c(tail(circColors2, -1), head(circColors2, 1))
```

## Raw data

### Case data

We read in the raw case data at the municipality level in Brazil. 

```{r}
bra2_case_data <- readRDS(file=bra2_case_data_path)
```

Let's get the state/city to city id mapping, so we can add it to other sets of data for ease of data management later on. 

```{r}
city_cityid_mapping <- bra2_case_data %>%
  select(city_id, state, city) %>%
  unique()
```

### Spatial data 

Here, we read in and format the spatial polygons data for Brazil. 

```{r}
bra1 <- readRDS(bra1_path) %>%
  mutate(x=st_coordinates(st_centroid(geometry))[,1], y=st_coordinates(st_centroid(geometry))[,2])

bra2 <- readRDS(bra2_path) %>%
  left_join(city_cityid_mapping, by=c("state", "city"))
```

### Population data

Here, we read in the population data at the municipality level in Brazil. 

```{r}
bra2_pop_data <- readRDS(file=bra2_pop_data_path) %>%
  left_join(city_cityid_mapping, by=c("state", "city"))
```

Let's also add population density data at the municipality level, which will be used extensively in the following analyses. 

```{r}
bra2_pop_data %<>%
  left_join(st_drop_geometry(bra2), by=c("state", "city", "city_id")) %>%
  mutate(den_km2=n/area_m2*10^6)
```

## Processed summary data 

### Climate data 

We read in the precipitation data at the municipality level in Brazil. 

```{r}
load(bra2_entire_period_prec_summary_data_path) #region_basic_summary_data
bra2_entire_period_prec_summary_data <- region_basic_summary_data
rm(region_basic_summary_data)
```

Next, we read in the correlation data for a typical year between log-incidence and precipitation.  

```{r}
load(bra2_log_incidence_prec_corr_data_path) #region_typical_year_corr_data
bra2_log_incidence_prec_corr_data <- region_typical_year_corr_data
rm(region_typical_year_corr_data)
```

### Case data

We read in the typical year data for the log-incidence in each municipality. 

```{r}
load(bra2_log_incidence_typical_year_summary_data_path) #region_typical_year_cases_data
bra2_log_incidence_typical_year_summary_data <- region_typical_year_cases_data
rm(region_typical_year_cases_data)
```

### Index P data

Here, we read in the Index P summary data at the municipality level in Brazil. 

```{r}
load(bra2_indexP_summary_data_path) #region_basic_summary_data
bra2_indexP_summary_data <- region_basic_summary_data
rm(region_basic_summary_data)
```

### Index P/case data

We read in the correlation summary data between Index P and log-incidence during a typical year. 

```{r}
load(bra2_typical_year_corr_log_incidence_data_path) #region_typical_year_corr_data
bra2_typical_year_corr_log_incidence_data <- region_typical_year_corr_data
rm(region_typical_year_corr_data)
```

## Accessory plotting functions

`plot_12_month_donut` plots the 12-month doughnut chart to be used for legends. It takes as arguments the colours, the associated values and the value at which the doughnut chart should be centred. 

```{r}
plot_12_month_donut <- function(colors, values, center) {
  # centre colours
  center_pos <- which(values==center)
  if (center_pos!=1) {
    values <- c(tail(values, -(center_pos-1)), head(values, (center_pos-1)))
    colors <- c(tail(colors, -(center_pos-1)), head(colors, (center_pos-1)))
  }

  # create the fictional data
  MONTHS_IN_YEAR <- 12
  data <- data.frame(category=values, count=rep(1, MONTHS_IN_YEAR))
  data$category <- factor(data$category, levels=values)
  data$fraction <- data$count/sum(data$count)
  data$ymax <- cumsum(data$fraction)
  data$ymin <- c(0, head(data$ymax, n=-1))
  data$labelPosition <- (data$ymax+data$ymin)/2
  data$label <- values

  # plot doughnut
  donut_plot <- ggplot(data, aes(ymax=ymax, ymin=ymin, xmax=4, xmin=3, fill=category), color="grey") +
    geom_rect() +
    geom_label(x=3.5, aes(y=labelPosition, label=label), fill="white", size=6) +
    geom_text(x=2, y=0, label="Phase\nshift", size=10) +
    scale_fill_manual(breaks=values, values=colors, guide="none") +
    coord_polar(theta="y") +
    xlim(c(2, 4)) +
    theme_void()
  return(donut_plot)
}

# examples
#plot_12_month_donut(colors=circColors, values=1:12, center=1)
#plot_12_month_donut(colors=circColors2, values=(-5):6, center=0)
```

## Index P and Case Burden

### DENV incidence

Let's compute summary statistics for DENV incidence in each municipality. 

```{r}
bra2_case_burden <- bra2_case_data %>% 
  filter(!grepl(pattern="Município ignorado", x=city, fixed=TRUE)) %>%
  group_by(city_id, state, city, year) %>%
  summarise(cases=sum(cases, na.rm=FALSE), .groups="keep") %>% 
  left_join(bra2_pop_data, by=c("state", "city", "year", "city_id")) %>%
  mutate(incidence=cases/n) %>% 
  as.data.frame() %>%
  group_by(city_id, state, city) %>%
  summarise(total_cases=sum(cases, na.rm=FALSE),
            mean_yearly_cases=mean(cases, na.rm=FALSE),
            median_yearly_cases=median(cases, na.rm=FALSE),
            mean_yearly_incidence=mean(incidence, na.rm=FALSE)*10^5, 
            median_yearly_incidence=median(incidence, na.rm=FALSE)*10^5, 
            years_with_cases=sum(incidence!=0, na.rm=FALSE), .groups="keep") %>%
  as.data.frame()
```

### Index P

Let's compute some summary measures of Index P. 

```{r}
# summaries that pertain to the entire period
retrieve_entire_period_summary <- function(summary_measure_id, summary_measure_name) {
  entire_period_summary <- bra2_indexP_summary_data[[summary_measure_id]] %>%
    as.data.frame() %>%
    gather(key="city_id", value=value) %>% 
    setNames(c("city_id", summary_measure_name))
}
input <- data.frame(summary_measure_id=c("entire_mean", "entire_median", "entire_sd", "mean_AUC", "median_AUC", "sd_AUC"), 
                    summary_measure_name=c("mean_cum", "median_cum", "sd_cum", "mean_auc", "median_auc", "sd_auc"))
entire_period_summaries <- Reduce(function(x, y) left_join(x, y, by="city_id"), 
                                  apply(input, 1, function(x) retrieve_entire_period_summary(x[1], x[2])))

# summaries that pertain to each year
retrieve_yearly_summary <- function(summary_measure_id, summary_measure_name, my_function) {
  yearly_summary <- bra2_indexP_summary_data[[summary_measure_id]] %>%
    as.data.frame() %>%
    gather(key="city_id", value=value, 2:ncol(.)) %>% 
    group_by(city_id)
  if (summary_measure_name=="typical_year_auc") yearly_summary %<>% summarise(value=my_function(x=month, y=value), 
                                                                              .groups="keep")
  else if (summary_measure_name=="typical_year_months_above_1") yearly_summary %<>% summarise(value=sum(value>1), 
                                                                                              .groups="keep")
  else if (summary_measure_name=="typical_year_months_above_0.5") yearly_summary %<>% summarise(value=sum(value>0.5), 
                                                                                              .groups="keep")
  else yearly_summary %<>% summarise(value=my_function(value), .groups="keep")
  yearly_summary %<>% setNames(c("city_id", summary_measure_name)) %>% as.data.frame()
  return(yearly_summary)
}
input <- data.frame(summary_measure_id=c("yearly_median", "median_AUC_per_year", "typical_year_mean", 
                                         "typical_year_mean", "typical_year_mean", "typical_year_median", 
                                         "typical_year_mean", "typical_year_mean"), 
                    summary_measure_name=c("mean_yearly_cum", "mean_yearly_auc", "typical_year_mean", 
                                           "typical_year_max", "typical_year_auc", "typical_year_median", 
                                           "typical_year_months_above_1", "typical_year_months_above_0.5"))
my_functions=c(mean, mean, mean, max, trapz, median, NULL)
yearly_summaries <- Reduce(function(x, y) left_join(x, y, by="city_id"), 
                           lapply(1:nrow(input), function(x) retrieve_yearly_summary(input[x, 1], input[x, 2], 
                                                                                     my_functions[[x]])))

# combine both summaries 
selected_indexP_summaries <- left_join(entire_period_summaries, yearly_summaries, by="city_id")
```

## Typical year correlation

In this section, we look at the correlation between Index P and incidence during a typical year and try to explain the variation that we observe assuming that higher correlation indicates that Index P is more reflective of the reported incidence dynamics. 

Let's start by calculating some correlation summary statistics. 

```{r}
typical_year_corr_no_lag_data <- bra2_typical_year_corr_log_incidence_data %>% filter(shift==0)
typical_year_corr_lag_data <- bra2_typical_year_corr_log_incidence_data %>% 
  group_by(city_id, state, city) %>% 
  summarise(opt_lag=ifelse(any(is.na(median)), NA, shift[which(median==max(median, na.rm=FALSE))]), 
            opt_median=ifelse(all(is.na(median)), NA, max(median, na.rm=TRUE)), .groups="keep") %>%
  as.data.frame() 
typical_year_corr_summary_data <- typical_year_corr_lag_data %>% 
  left_join(typical_year_corr_no_lag_data, by=c("state", "city", "city_id"))
```

## Figures

First, we plot the summary figures for correlation, phase-shifted correlation and the phase shift. 

```{r warning=FALSE}
plot_corr_complete_plots <- function(typical_year_corr_data, incidence_data, threshold, bra2, bra1, file_name) {
  no_lag_typical_year_corr_data <- typical_year_corr_data %>% 
    left_join(mutate(incidence_data, city_id=as.numeric(city_id)), by=c("city_id", "state", "city")) %>% 
    mutate(median=ifelse(mean_yearly_cases<threshold, NA, median)) %>% 
    mutate(opt_median=ifelse(mean_yearly_cases<threshold, NA, opt_median)) %>% 
    mutate(opt_lag=ifelse(mean_yearly_cases<threshold, NA, opt_lag)) %>% 
    left_join(mutate(bra2, city_id=as.numeric(city_id)), by=c("city_id", "state", "city")) %>% 
    st_as_sf()

  # spatial map of correlations
  corr_no_lag_plot <- ggplot() + 
    geom_sf(data=no_lag_typical_year_corr_data, 
            mapping=aes(geometry=geometry), fill="grey", colour="darkgrey", lwd=0.1) + 
    geom_sf(data=filter(no_lag_typical_year_corr_data, !is.na(median)), 
            mapping=aes(geometry=geometry, fill=median, color=median), lwd=0.1) + 
    geom_sf(data=bra1, 
            mapping=aes(geometry=geometry), fill=NA, color="black", lwd=0.1) + 
    scale_fill_gradientn(colours=rev(brewer.pal(11, "Spectral")), na.value="black", limits=c(-1, 1), 
                         breaks=round(seq(-1, 1, 0.5), 1), labels=c("-1.0", "-0.5", " 0.0", " 0.5", " 1.0"),
                         name=NULL) + 
    scale_color_gradientn(colours=rev(brewer.pal(11, "Spectral")), na.value="black", guide="none", limits=c(-1, 1)) + 
    scale_x_continuous(limits=c(-73, -36)) + 
    guides(fill=guide_colorbar(ticks.colour="black", frame.colour="black")) + 
    theme(panel.grid=element_blank(), axis.line=element_blank(), 
          axis.text=element_blank(), axis.title=element_blank(), 
          panel.background=element_blank(), axis.ticks=element_blank(), 
          legend.title=element_blank()) + 
    theme(legend.position="right", legend.title.align = 0.5, 
          legend.key.width=unit(0.25, "cm"), legend.key.height=unit(1.25, "cm")) + 
    theme(plot.margin=unit(c(0, 0.1, 0, 0), "cm"))
  
  # histogram of the distribution of possible correlations
  mean_corr <- no_lag_typical_year_corr_data %>% pull(median) %>% mean(na.rm=TRUE)
  corr_no_lag_histogram <- ggplot() + 
    geom_histogram(data=no_lag_typical_year_corr_data, 
                   mapping=aes(x=median), color="black", fill="white", binwidth=0.05) + 
    geom_vline(xintercept=mean_corr, linetype="dashed", colour="red") + 
    scale_x_continuous(limits=c(-1, 1), breaks=seq(-1, 1, 0.5), name=NULL) + 
    #labs(x="Correlation") + 
    theme_bw() + 
    theme(axis.title.y=element_blank(), axis.text.y=element_blank()) + 
    theme(plot.margin=unit(c(0, 0.1, 0.1, 0.5), "cm"))
  
  # lag and correlation 
  boxplot_typical_year_corr_data <- no_lag_typical_year_corr_data %>%
    st_drop_geometry() %>% 
    filter(!is.na(median)) %>% 
    group_by(opt_lag) %>%
    mutate(num_mun=length(city_id), percent_mun=length(city_id)/nrow(.))
  opt_lag_boxplot <- ggplot() + 
    geom_boxplot(data=mutate(boxplot_typical_year_corr_data, opt_lag=factor(opt_lag, levels=(-5):6)), 
                 mapping=aes(x=opt_lag, y=median, fill=percent_mun*100)) + 
    geom_hline(yintercept=mean_corr, linetype="dashed", color="red") + 
    geom_hline(yintercept=0, linetype="solid", color="black") + 
    labs(x="Phase shift (months)", y="Correlation", fill="Percentage of muncipalities") + 
    scale_y_continuous(limits=c(-1, 1), breaks=seq(-1, 1, 0.25)) + 
    # scale_fill_gradientn(colors=rep(levColors, each=3), 
    #                      guide=guide_colorbar(override.aes=list(fill=NA), title.position="top", frame.colour="black", ticks.colour="black"), 
    #                      breaks=c(0, 750, 1500), limits=c(0, 1500)) + 
    scale_fill_gradientn(colors=rep(levColors, each=3), 
                         guide=guide_colorbar(override.aes=list(fill=NA), title.position="top", frame.colour="black", ticks.colour="black"), 
                         breaks=seq(0, 50, 10), limits=c(0, 50), labels=paste(seq(0, 50, 10), "%", sep="")) + 
    theme_bw() + 
    theme(legend.position="bottom", legend.title.align=0.4, legend.direction="horizontal", 
          legend.key.width=unit(0.85, "cm"), legend.key.size=unit(0.5, "cm"), legend.key.height=unit(0.3, "cm"), 
          legend.title=element_text(size=9), legend.text=element_text(size=9))
    # theme(legend.position=c(0.83, 0.85), legend.key.size=unit(0.25, "cm"), legend.direction="horizontal", 
    #       legend.background=element_blank(), legend.key.width=unit(0.55, "cm"), 
    #       legend.text=element_text(size=7), legend.title=element_text(size=7), legend.title.align=0.4)
  
    # spatial map of correlations
  corr_opt_lag_plot <- ggplot() + 
    geom_sf(data=no_lag_typical_year_corr_data, 
            mapping=aes(geometry=geometry), fill="grey", colour="darkgrey", lwd=0.1) + 
    geom_sf(data=filter(no_lag_typical_year_corr_data, !is.na(opt_median)), 
            mapping=aes(geometry=geometry, fill=opt_median, color=opt_median), lwd=0.1) + 
    geom_sf(data=bra1, 
            mapping=aes(geometry=geometry), fill=NA, color="black", lwd=0.1) + 
    scale_fill_gradientn(colours=rev(brewer.pal(11, "Spectral")), na.value="black", limits=c(-1, 1), 
                         breaks=seq(-1.0, 1, 0.5), labels=c("-1.0", "-0.5", " 0.0", " 0.5", " 1.0"), 
                         name=NULL) + 
    scale_color_gradientn(colours=rev(brewer.pal(11, "Spectral")), na.value="black", guide="none", limits=c(-1, 1)) + 
    scale_x_continuous(limits=c(-73, -36)) + 
    guides(fill=guide_colorbar(ticks.colour="black", frame.colour="black")) + 
    theme(panel.grid=element_blank(), axis.line=element_blank(), 
          axis.text=element_blank(), axis.title=element_blank(), 
          panel.background=element_blank(), axis.ticks=element_blank(), 
          legend.title=element_blank()) + 
    theme(legend.position="right", legend.title.align = 0.5, 
          legend.key.width=unit(0.25, "cm"), legend.key.height=unit(1.25, "cm")) + 
    theme(plot.margin=unit(c(0, 0.1, 0, 0), "cm"))
  
  mean_opt_corr <- no_lag_typical_year_corr_data %>% pull(opt_median) %>% mean(na.rm=TRUE)
  corr_lag_histogram <- ggplot() + 
    geom_histogram(data=no_lag_typical_year_corr_data, 
                   mapping=aes(x=opt_median), color="black", fill="white", binwidth=0.05) + 
    geom_vline(xintercept=mean_opt_corr, linetype="dashed", colour="red") + 
    scale_x_continuous(limits=c(-1, 1), breaks=seq(-1, 1, 0.5), name=NULL) + 
    #labs(x="Phase shifted correlation") + 
    theme_bw() + 
    theme(axis.title.y=element_blank(), axis.text.y=element_blank()) + 
    theme(plot.margin=unit(c(0, 0.1, 0.1, 0.5), "cm"))
  
  mun_timing_map <- ggplot() + 
    geom_sf(data=no_lag_typical_year_corr_data, 
            mapping=aes(geometry=geometry), fill="lightgrey", colour="darkgrey", lwd=0.1) + 
    geom_sf(data=filter(no_lag_typical_year_corr_data, !is.na(opt_lag)), 
            mapping=aes(geometry=geometry, fill=factor(opt_lag), color=factor(opt_lag)), lwd=0) + 
    geom_sf(data=bra1, 
            mapping=aes(geometry=geometry), fill=NA, color="black", lwd=0.1) + 
    scale_fill_manual(breaks=(-5):6, values=circColors2, na.value="black", drop=FALSE, guide="none") +
    scale_color_manual(breaks=(-5):6, values=circColors2, na.value="black", guide="none") +
    scale_x_continuous(limits=c(-73, -36)) + 
    theme(panel.grid=element_blank(), axis.line=element_blank(), 
          axis.text=element_blank(), axis.title=element_blank(), 
          panel.background=element_blank(), axis.ticks=element_blank(), 
          legend.title=element_blank())
  
  month_donut <- plot_12_month_donut(colors=circColors2, values=(-5):6, center=0) + 
    theme(plot.margin=unit(c(0, 0, 0, 0), "cm"))
  
  no_lag_corr_title <- ggdraw() + draw_label("Correlation") + theme(plot.margin=margin(0, 0, 0, 0))
  lag_corr_title <- ggdraw() + draw_label("Phase-shifted correlation") + theme(plot.margin=margin(0, 0, 0, 0))
  
  pB <- cowplot::plot_grid(corr_lag_histogram, corr_opt_lag_plot, ncol=1, rel_heights=c(1, 2))
  pB <- cowplot::plot_grid(lag_corr_title, pB, ncol=1, rel_heights=c(0.1, 1))
  pA <- cowplot::plot_grid(corr_no_lag_histogram, corr_no_lag_plot, ncol=1, rel_heights=c(1, 2))
  pA <- cowplot::plot_grid(no_lag_corr_title, pA, ncol=1, rel_heights=c(0.1, 1))
  pAB <- cowplot::plot_grid(pA, pB, labels=c("A", "B"), nrow=1)
  pC <- cowplot::plot_grid(opt_lag_boxplot, NULL, ncol=1, rel_heights=c(1.5, 0.25))
  pCD <- cowplot::plot_grid(pC, mun_timing_map, nrow=1, labels=c("C", "D"))
  p <- cowplot::plot_grid(pAB, pCD, ncol=1, rel_heights=c(1.5, 1))
  scale_factor=0.75
  pdf(file.path(SUMMARY_PLOTS_FOLDER, file_name), w=8.27, h=11.69*scale_factor)
  print(p)
  a <- dev.off()
  
  pdf(file.path(SUMMARY_PLOTS_FOLDER, "donut_plot.pdf"), w=5, h=5)
  print(month_donut)
  a <- dev.off()
}
plot_corr_complete_plots(typical_year_corr_data=typical_year_corr_summary_data, 
                         incidence_data=bra2_case_burden, 
                         threshold=12, 
                         bra2=bra2, bra1=bra1, file_name="typical_year_corr_complete_plots.pdf")
```

Next, we define a function to compare the typical year dynamics of Index P, precipitation and incidence. All three variables with be normalised between 0 and 1. 

```{r}
plot_combined_typical_year_timeseries <- function(regions, 
                                                  indexP_summary_data=bra2_indexP_summary_data$typical_year_mean,
                                                  log_incidence_typical_year_summary_data=bra2_log_incidence_typical_year_summary_data, 
                                                  prec_typical_year_summary_data=bra2_entire_period_prec_summary_data$typical_year) {
  complete_typical_year_data_list <- list()
  for (ii in 1:nrow(regions)) {
    this_city_id <- regions[ii, "city_id"]
    this_state <- regions[ii, "state"]
    this_city <- regions[ii, "city"]
    
    # index P typical year 
    city_id_pos <- which(colnames(indexP_summary_data)==as.character(this_city_id))
    indexP_typical_year <- indexP_summary_data[, c(1, city_id_pos)] %>% 
      as.data.frame() %>%
      setNames(c("month", "unnormalised_indexP")) %>% 
      mutate(normalised_indexP=(unnormalised_indexP-min(unnormalised_indexP))/(max(unnormalised_indexP)-min(unnormalised_indexP)))
    
    # log-incidence typical year
    log_incidence_typical_year <- log_incidence_typical_year_summary_data %>% 
      filter(city_id==this_city_id) %>% 
      select(month, cases) %>% 
      setNames(c("month", "unnormalised_log_incidence")) %>%
      mutate(normalised_log_incidence=(unnormalised_log_incidence-min(unnormalised_log_incidence))/
               (max(unnormalised_log_incidence)-min(unnormalised_log_incidence)))
    
    # precipitation typical year 
    city_id_pos <- which(colnames(prec_typical_year_summary_data)==as.character(this_city_id))
    prec_typical_year <- prec_typical_year_summary_data[, c(1, city_id_pos)] %>% 
      as.data.frame() %>%
      setNames(c("month", "unnormalised_prec")) %>% 
      mutate(normalised_prec=(unnormalised_prec-min(unnormalised_prec))/(max(unnormalised_prec)-min(unnormalised_prec)))
    
    complete_typical_year_data <- indexP_typical_year %>% 
      left_join(log_incidence_typical_year, by="month") %>%
      left_join(prec_typical_year, by="month") %>% 
      mutate(city_id=this_city_id, state=this_state, city=this_city) %>% 
      select(city_id, state, city, month, normalised_indexP, normalised_log_incidence, normalised_prec)
    complete_typical_year_data_list[[ii]] <- complete_typical_year_data
  }
  complete_typical_year_data <- do.call(rbind, complete_typical_year_data_list) %>% 
    gather(key="variable", value="value", c("normalised_indexP", "normalised_log_incidence", "normalised_prec")) %>% 
    mutate(city_name=paste0(city, ", ", state)) %>% 
    mutate(city_name=factor(city_name, levels=pull(mutate(regions, city_name=paste0(city, ", ", state)), city_name))) %>% 
    mutate(month=factor(month, levels=c(7:12, 1:6)))

  
  p <- ggplot() + 
    geom_line(data=complete_typical_year_data, 
              mapping=aes(x=month, y=value, group=variable, color=variable)) +
    labs(y="Normalised value", x="Month") + 
    scale_y_continuous(limits=c(0, 1), breaks=seq(0, 1, 0.25)) + 
    scale_color_manual(breaks= c("normalised_indexP", "normalised_log_incidence", "normalised_prec"), 
                       values=c("#E69F00", "#0072B2", "#009E73"), 
                       labels=c("Index P", "Incidence (log)", "Rainfall"), name=NULL) + 
    theme_bw() + 
    guides(color=guide_legend(override.aes = list(size=3))) + 
    theme(legend.position="bottom") + 
    theme(legend.box.margin=margin(-10,-10,-10,-10)) + 
    facet_wrap(~city_name, nrow=2) 
  return(p)
}
```

Finally, let's plot the collection of figures that suggests that the omission of rainfall may explain the aberrant phase shifts. 

```{r warning=FALSE}
plot_phase_shift_prec_complete_plots <- function(typical_year_corr_data, incidence_data, threshold, precipitation_data, indexP_summary_data, 
                                                 typical_year_prec_corr_data, bra2=bra2, bra1=bra1, file_name) {
  typical_year_corr_data %<>%    
    left_join(mutate(incidence_data, city_id=as.numeric(city_id)), by=c("city_id", "state", "city")) %>% 
    mutate(median=ifelse(mean_yearly_cases<threshold, NA, median)) %>% 
    mutate(opt_median=ifelse(mean_yearly_cases<threshold, NA, opt_median)) %>% 
    mutate(opt_lag=ifelse(mean_yearly_cases<threshold, NA, opt_lag)) %>% 
    right_join(mutate(bra2, city_id=as.numeric(city_id)), by=c("city_id", "state", "city")) %>%
    mutate(opt_lag=factor(opt_lag, levels=(-5):6))
  
  selected_typical_year_corr_data <- typical_year_corr_data %>%
    filter(opt_lag %in% c((-5):(-1), 4:6)) %>% 
    left_join(typical_year_prec_corr_data, by=c("city_id", "state", "city"))
  
  focal_city_ids <- c(160030, 140030, 210540, 150680, 160040, 140017, 210990, 150210)
  regions <- bra2 %>% 
    st_drop_geometry() %>% 
    select(city_id, state, city) %>% 
    filter(city_id %in% focal_city_ids)
  regions <- regions[match(focal_city_ids, regions$city_id), ]
  
  
  map_text <- bra2 %>% 
    filter(city_id %in% focal_city_ids) %>% 
    mutate(x=st_coordinates(st_centroid(geometry))[,1], y=st_coordinates(st_centroid(geometry))[,2])
    
  mun_timing_map <- ggplot() + 
    geom_sf(data=selected_typical_year_corr_data,  
            mapping=aes(geometry=geometry, fill=opt_lag, color=opt_lag), lwd=0) + 
    geom_sf(data=bra1, 
            mapping=aes(geometry=geometry), fill=NA, color="black", lwd=0.1) + 
    geom_point(data=map_text, mapping=aes(x=x, y=y), color="red") + 
    geom_label_repel(data=filter(map_text, city_id %in% c(160030, 210540, 140030, 140017)), 
                     mapping=aes(x=x, y=y, label=city), 
                     size=1.75, min.segment.length=0, nudge_y=5, label.padding=unit(0.1, "cm"), box.padding=unit(0.1, "cm")) + 
    geom_label_repel(data=filter(map_text, city_id %in% c(150680, 210990, 150210)), 
                     mapping=aes(x=x, y=y, label=city), 
                     size=1.75, min.segment.length=0, nudge_y=-4, 
                     label.padding=unit(0.1, "cm"), box.padding=unit(0.1, "cm")) + 
    geom_label_repel(data=filter(map_text, city_id %in% c(160040)), 
                     mapping=aes(x=x, y=y, label=city), 
                     size=1.75, min.segment.length=0, label.padding=unit(0.1, "cm"), box.padding=unit(0.1, "cm")) + 
    scale_fill_manual(breaks=(-5):6, values=circColors2, na.value="black", drop=FALSE, guide="none") +
    scale_color_manual(breaks=(-5):6, values=circColors2, na.value="black", guide="none") +
    scale_x_continuous(limits=c(-73, -36)) + 
    theme(panel.grid=element_blank(), axis.line=element_blank(), 
          axis.text=element_blank(), axis.title=element_blank(), 
          panel.background=element_blank(), axis.ticks=element_blank(), 
          legend.title=element_blank()) + 
    theme(plot.margin=unit(c(0, 0, 0, 0), "cm"))
  
  # precipitation corr vs. phase-adjusted corr
  corr_vs_corr_scatterplot <- ggplot() + 
    geom_point(data=selected_typical_year_corr_data,  
               mapping=aes(x=corr, y=median, color=opt_lag), size=2) + 
    geom_abline(intercept=0, slope=1, linetype="dashed") + 
    labs(x="Precipitation-incidence correlation", y="Index P-incidence correlation") + 
    scale_color_manual(breaks=(-5):6, values=circColors2, na.value="black", drop=FALSE, guide="none") +
    scale_x_continuous(limits=c(-1, 1), breaks=seq(-1, 1, 0.25)) + 
    scale_y_continuous(limits=c(-1, 1), breaks=seq(-1, 1, 0.25)) + 
    theme_bw()
  
  # precipitation, index P and incidence dynamics examples
  combined_timeseries <- plot_combined_typical_year_timeseries(regions=regions)
  
  pAB <- cowplot::plot_grid(mun_timing_map, corr_vs_corr_scatterplot, nrow=1, labels=c("A", "B"))
  pABC <- cowplot::plot_grid(pAB, combined_timeseries, ncol=1, labels=c("", "C"))
  scale_factor <- 0.6
  pdf(file.path(SUMMARY_PLOTS_FOLDER, file_name), w=8.27, h=11.69*scale_factor)
  print(pABC)
  a <- dev.off()
}
plot_phase_shift_prec_complete_plots(typical_year_corr_data=typical_year_corr_summary_data, 
                                     indexP_summary_data=selected_indexP_summaries, 
                                     threshold=12, 
                                     incidence_data=bra2_case_burden, bra2=bra2, bra1=bra1,
                                     precipitation_data=bra2_entire_period_prec_summary_data, 
                                     typical_year_prec_corr_data=bra2_log_incidence_prec_corr_data, 
                                     file_name="typical_year_phase_shift_prec_complete_plots.pdf")
```
